requirements.txt
pandas>=1.3.0
numpy>=1.21.0
pandas-datareader>=0.10.0
scikit-learn>=1.0.0
plotly>=5.0.0
matplotlib>=3.4.0
wordcloud>=1.8.0
tinydb>=4.5.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.0.0
joblib>=1.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=3.0.0
